credit_approval_system/requirements.txt
django>=4.0
djangorestframework
psycopg2-binary
pandas
openpyxl
celery
redis
python-dotenv